app/utils/preprocessing.py
import numpy as np
from PIL import Image


def preprocess_image(file):
    image = Image.open(file).convert("RGB")
    image = image.resize((224, 224))
    img_array = np.array(image).astype(np.float32) / 255.0
    img_array = np.transpose(img_array, (2, 0, 1))
    img_array = np.expand_dims(img_array, axis=0)
    
    return img_array