requirements.txt
fastapi
uvicorn[standard]
numpy
pillow
tritonclient[http]
python-multipart
prometheus-fastapi-instrumentator
loguru